+++ scripts/tests/forest_cli_check.sh
@@ -35,7 +35,7 @@ pushd "$(mktemp --directory)"
 
     : verify that we are byte-for-byte identical with filops
     zstd -d filops_*.car.zst
-    "$FOREST_CLI_PATH" archive export filops_*.car -o exported_snapshot.car.zst
+    "$FOREST_TOOL_PATH" archive export filops_*.car -o exported_snapshot.car.zst
     zstd -d exported_snapshot.car.zst
     cmp --silent filops_*.car exported_snapshot.car
 
@@ -44,7 +44,7 @@ pushd "$(mktemp --directory)"
 
     : verify that diff exports contain the expected number of state roots
     EPOCH=$(forest_query_epoch exported_snapshot.car.zst)
-    "$FOREST_CLI_PATH" archive export --epoch $((EPOCH-1100)) --depth 900 --output-path base_snapshot.forest.car.zst exported_snapshot.car.zst
+    "$FOREST_TOOL_PATH" archive export --epoch $((EPOCH-1100)) --depth 900 --output-path base_snapshot.forest.car.zst exported_snapshot.car.zst
 
     BASE_EPOCH=$(forest_query_epoch base_snapshot.forest.car.zst)
     assert_eq "$BASE_EPOCH" $((EPOCH-1100))
@@ -53,7 +53,7 @@ pushd "$(mktemp --directory)"
     #BASE_STATE_ROOTS=$(forest_query_state_roots base_snapshot.forest.car.zst)
     #assert_eq "$BASE_STATE_ROOTS" 900
 
-    "$FOREST_CLI_PATH" archive export --diff "$BASE_EPOCH" -o diff_snapshot.forest.car.zst exported_snapshot.car.zst
+    "$FOREST_TOOL_PATH" archive export --diff "$BASE_EPOCH" -o diff_snapshot.forest.car.zst exported_snapshot.car.zst
     # This assertion is not true in the presence of null tipsets
     #DIFF_STATE_ROOTS=$(forest_query_state_roots diff_snapshot.forest.car.zst)
     #assert_eq "$DIFF_STATE_ROOTS" 1100
@@ -86,7 +86,7 @@ pushd "$(mktemp --directory)"
 
     : : check that it contains at least one expected checkpoint
     # If calibnet is reset or the checkpoint interval is changed, this check has to be updated
-    "$FOREST_CLI_PATH" archive checkpoints "$validate_me" | grep bafy2bzaceatx7tlwdhez6vyias5qlhaxa54vjftigbuqzfsmdqduc6jdiclzc
+    "$FOREST_TOOL_PATH" archive checkpoints "$validate_me" | grep bafy2bzaceatx7tlwdhez6vyias5qlhaxa54vjftigbuqzfsmdqduc6jdiclzc
 rm -- *
 popd